requirements.txt
Django>=4.0,<5.0
graphene-django>=3.0,<4.0
django-filter>=23.0,<24.0